cambios

+++ src/main/java/ms/hispam/budget/util/ExcelService.java
@@ -115,7 +115,7 @@ public class ExcelService {
                Cell finalPdataCell = pdataCell;
                int finalJ = j;
                dataBase.getData().get(i).getComponents().stream().filter(r->r.getComponent().equalsIgnoreCase(dataBase.getComponents().get(finalJ).getComponent())).findFirst()
-                       .ifPresentOrElse(r-> finalPdataCell.setCellValue(r.getAmount()),
+                       .ifPresentOrElse(r-> finalPdataCell.setCellValue(r.getAmount().doubleValue()),
                        ()->finalPdataCell.setCellValue(0));
                starDetail++;
            }
@@ -124,7 +124,7 @@ public class ExcelService {
                Cell finalPdataCell = pdataCell;
                int finalK= k;
                dataBase.getData().get(i).getComponents().stream().filter(r->r.getComponent().equalsIgnoreCase(dataBase.getNominas().get(finalK).getCodeNomina())).findFirst()
-                       .ifPresentOrElse(r-> finalPdataCell.setCellValue(r.getAmount()),
+                       .ifPresentOrElse(r-> finalPdataCell.setCellValue(r.getAmount().doubleValue()),
                                ()->finalPdataCell.setCellValue(0));
                starDetail++;
            }
